requirements.txt
netifaces
python-telegram-bot